# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Wabiz CRM API",
        "description": "Wabiz CRM - Multi-tenant WhatsApp CRM SaaS Platform API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Development Server"
        }
    ],
    "paths": {
...
        "/billing/plans": {
            "get": {
                "tags": [
                    "Billing"
                ],
                "summary": "Get all subscription plans",
                "description": "Retrieve all active subscription plans with pricing, limits, and features. This is a public endpoint.",
                "operationId": "26124f99e59e128359125c5424764dbc",
                "responses": {
                    "200": {
                        "description": "Plans fetched successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Plans fetched successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/SubscriptionPlan"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/billing/subscription": {
            "get": {
                "tags": [
                    "Billing"
                ],
                "summary": "Get current subscription",
                "description": "Retrieve the authenticated user's active subscription with plan details, remaining days, and usage stats.",
                "operationId": "1e86aac3b7498811e9646890fe86e799",
                "responses": {
                    "200": {
                        "description": "Subscription fetched successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "properties": {
                                                "subscription": {
                                                    "$ref": "#/components/schemas/UserSubscription"
                                                },
                                                "usage": {
                                                    "properties": {
                                                        "leads": {
                                                            "properties": {
                                                                "current": {
                                                                    "type": "integer",
                                                                    "example": 25
                                                                },
                                                                "limit": {
                                                                    "type": "integer",
                                                                    "example": 50
                                                                }
                                                            },
                                                            "type": "object"
                                                        },
                                                        "messages": {
                                                            "properties": {
                                                                "current": {
                                                                    "type": "integer",
                                                                    "example": 40
                                                                },
                                                                "limit": {
                                                                    "type": "integer",
                                                                    "example": 100
                                                                }
                                                            },
                                                            "type": "object"
                                                        },
                                                        "orders": {
                                                            "properties": {
                                                                "current": {
                                                                    "type": "integer",
                                                                    "example": 5
                                                                },
                                                                "limit": {
                                                                    "type": "integer",
                                                                    "example": 10
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No active subscription"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "sanctumAuth": []
                    }
                ]
            }
        },
        "/billing/subscribe": {
            "post": {
                "tags": [
                    "Billing"
                ],
                "summary": "Subscribe to a plan",
                "description": "Create a new subscription for the authenticated user. For free plans, subscription activates immediately. For paid plans, provide a payment ID from the gateway.",
                "operationId": "e54f227aa47c4933a8e80699684b7346",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "plan_id",
                                    "gateway"
                                ],
                                "properties": {
                                    "plan_id": {
                                        "description": "Subscription plan ID",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "gateway": {
                                        "type": "string",
                                        "example": "razorpay",
                                        "enum": [
                                            "razorpay",
                                            "stripe",
                                            "manual"
                                        ]
                                    },
                                    "payment_id": {
                                        "description": "Payment/order ID from the gateway",
                                        "type": "string",
                                        "example": "pay_ABC123",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Subscription created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/UserSubscription"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Already has active subscription"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "sanctumAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "SubscriptionPlan": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Pro"
                    },
                    "slug": {
                        "type": "string",
                        "example": "pro"
                    },
                    "description": {
                        "type": "string",
                        "example": "Professional plan with higher limits",
                        "nullable": true
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 999
                    },
                    "billing_cycle": {
                        "type": "string",
                        "example": "monthly",
                        "enum": [
                            "monthly",
                            "yearly"
                        ]
                    },
                    "max_leads": {
                        "type": "integer",
                        "example": 500
                    },
                    "max_messages": {
                        "type": "integer",
                        "example": 5000
                    },
                    "max_orders": {
                        "type": "integer",
                        "example": 200
                    },
                    "features": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "example": "active",
                        "enum": [
                            "active",
                            "inactive"
                        ]
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctumAuth": {
                "type": "http",
                "bearerFormat": "token",
                "scheme": "bearer"
            }
        }
    },
...
}